Migrates project to PostgreSQL database
Adapts the application to use PostgreSQL instead of SQLite.

This change addresses issues with SQLite's compatibility and scalability in the production environment, as well as inconsistencies between development and production databases.

It includes updates to database connection, queries and data types in multiple scripts. Also includes project documentation updates to reflect architectural changes from stage 1 to stage 2, and future plans.